+++ packages/config/src/configs/moonbeam.ts
@@ -48,7 +48,7 @@ import {
   centrifuge,
   darwinia,
   equilibrium,
-  hydraDX,
+  hydration,
   interlay,
   mantaParachain,
   moonbeam,
@@ -117,7 +117,7 @@ export const moonbeamConfig = new ChainConfig({
       asset: glmr,
       balance: BalanceBuilder().substrate().system().account(),
       contract: ContractBuilder().Xtokens().transfer(),
-      destination: hydraDX,
+      destination: hydration,
       destinationFee: {
         amount: 0.05,
         asset: glmr,
@@ -466,7 +466,7 @@ export const moonbeamConfig = new ChainConfig({
       contract: ContractBuilder().Xtokens().transfer(),
       destination: polkadotAssetHub,
       destinationFee: {
-        amount: 0.7,
+        amount: 0.2,
         asset: usdt,
         balance: BalanceBuilder().substrate().assets().account(),
       },
@@ -481,7 +481,7 @@ export const moonbeamConfig = new ChainConfig({
       contract: ContractBuilder().Xtokens().transfer(),
       destination: polkadotAssetHub,
       destinationFee: {
-        amount: 0.7,
+        amount: 0.2,
         asset: usdc,
         balance: BalanceBuilder().substrate().assets().account(),
       },
@@ -496,7 +496,7 @@ export const moonbeamConfig = new ChainConfig({
       contract: ContractBuilder().Xtokens().transferMultiCurrencies(),
       destination: polkadotAssetHub,
       destinationFee: {
-        amount: 0.7,
+        amount: 0.2,
         asset: usdt,
         balance: BalanceBuilder().substrate().assets().account(),
       },
@@ -539,7 +539,7 @@ export const moonbeamConfig = new ChainConfig({
       asset: hdx,
       balance: BalanceBuilder().substrate().assets().account(),
       contract: ContractBuilder().Xtokens().transfer(),
-      destination: hydraDX,
+      destination: hydration,
       destinationFee: {
         amount: 0.6,
         asset: hdx,
@@ -554,7 +554,7 @@ export const moonbeamConfig = new ChainConfig({
       asset: dai,
       balance: BalanceBuilder().evm().erc20(),
       contract: ContractBuilder().Xtokens().transfer(),
-      destination: hydraDX,
+      destination: hydration,
       destinationFee: {
         amount: 0.004,
         asset: dai,
@@ -569,7 +569,7 @@ export const moonbeamConfig = new ChainConfig({
       asset: usdcwh,
       balance: BalanceBuilder().evm().erc20(),
       contract: ContractBuilder().Xtokens().transfer(),
-      destination: hydraDX,
+      destination: hydration,
       destinationFee: {
         amount: 0.004,
         asset: usdcwh,
@@ -599,7 +599,7 @@ export const moonbeamConfig = new ChainConfig({
       asset: usdtwh,
       balance: BalanceBuilder().evm().erc20(),
       contract: ContractBuilder().Xtokens().transfer(),
-      destination: hydraDX,
+      destination: hydration,
       destinationFee: {
         amount: 0.004,
         asset: usdtwh,
@@ -689,7 +689,7 @@ export const moonbeamConfig = new ChainConfig({
       asset: wbtc,
       balance: BalanceBuilder().evm().erc20(),
       contract: ContractBuilder().Xtokens().transfer(),
-      destination: hydraDX,
+      destination: hydration,
       destinationFee: {
         amount: 0.0000001,
         asset: wbtc,
@@ -704,7 +704,7 @@ export const moonbeamConfig = new ChainConfig({
       asset: weth,
       balance: BalanceBuilder().evm().erc20(),
       contract: ContractBuilder().Xtokens().transfer(),
-      destination: hydraDX,
+      destination: hydration,
       destinationFee: {
         amount: 0.000002,
         asset: weth,